DS-69: Add Button component
Based on the govuk component, with custom styling. Includes primary,
secondary, and warning versions.

Secondary and warning versions are not available in govuk@2.7.0 so
an additional version (govuk@2.12.0) has been added.

The build script has been updated to handle multiple versions of the
same source repository. The -s (--sources) flag can now be passed with
an optional version number to build a specific source.